# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "description": "This is an example API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/branch": {
            "get": {
                "tags": [
                    "Branch"
                ],
                "summary": "Get all Branch",
                "operationId": "7a60d8d8fd907b42bfb98cf328e9b958",
                "responses": {
                    "200": {
                        "description": "Everything OK"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/branch/getBranch": {
            "get": {
                "tags": [
                    "Branch"
                ],
                "summary": "Get  Branch",
                "operationId": "7017ddb323b2fe5d3eea2154fe59f4e2",
                "responses": {
                    "200": {
                        "description": "Everything OK"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/branch/store": {
            "post": {
                "tags": [
                    "Branch"
                ],
                "description": "",
                "operationId": "871c4487e83a3076748e73aabe9932c0",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "restaurant_id": {
                                                "type": "integer"
                                            },
                                            "name_ar": {
                                                "type": "string"
                                            },
                                            "name_en": {
                                                "type": "string"
                                            },
                                            "description_ar": {
                                                "type": "string"
                                            },
                                            "description_en": {
                                                "type": "string"
                                            },
                                            "address_ar": {
                                                "type": "string"
                                            },
                                            "address_en": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            },
                                            "mobile": {
                                                "type": "string"
                                            },
                                            "QR": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "restaurant_id": "restaurant_id",
                                    "name_ar": "name_ar",
                                    "name_en": "name_en",
                                    "description_ar": "description_ar",
                                    "description_en": "description_en",
                                    "address_ar": "address_ar",
                                    "address_en": "address_en",
                                    "phone": "phone",
                                    "mobile": "mobile",
                                    "QR": "QR"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "restaurant_id": {
                                            "type": "integr",
                                            "example": "restaurant_id"
                                        },
                                        "name_ar": {
                                            "type": "string",
                                            "example": "name_ar"
                                        },
                                        "name_en": {
                                            "type": "string",
                                            "example": "name_en"
                                        },
                                        "description_ar": {
                                            "type": "string",
                                            "example": "description_ar"
                                        },
                                        "description_en": {
                                            "type": "string",
                                            "example": "description_en"
                                        },
                                        "address_ar": {
                                            "type": "string",
                                            "example": "address_ar"
                                        },
                                        "address_en": {
                                            "type": "string",
                                            "example": "address_en"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "phone"
                                        },
                                        "mobile": {
                                            "type": "string",
                                            "example": "mobile"
                                        },
                                        "QR": {
                                            "type": "string",
                                            "example": "QR"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "CredentialsError = 1;Required = 100;MaxLength255 = 101;Unique = 103;"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/branch/show/{id}": {
            "get": {
                "tags": [
                    "Branch"
                ],
                "summary": "Get  Branch",
                "operationId": "628ee12c9268b45354c1b14dfa3303df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything OK"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/branch/delete/{id}": {
            "delete": {
                "tags": [
                    "Branch"
                ],
                "summary": "delete  Branch",
                "operationId": "9efbe728b8635ca28859ae0a4b09fc1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything OK"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}